CorrectInputTest replaced with Selenide version
